resort and split code
